streamlit-multipage-app-example-master/requirements.txt
absl-py==1.4.0
altair==4.2.2
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
blinker==1.5
build==0.10.0
cachetools==5.3.0
category-encoders==2.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
check-manifest==0.49
click==8.1.3
cmake==3.25.2
colorama==0.4.6
comm==0.1.2
comtypes==1.1.14
contourpy==1.0.7
coverage==7.1.0
cryptography==39.0.1
cycler==0.10.0
Cython==0.29.34
dash==2.9.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
easydict==1.10
easyocr==1.6.2
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
filelock==3.9.0
filterpy==1.4.5
flake8==6.0.0
Flask==2.2.3
flatbuffers==23.1.21
fonttools==4.38.0
future==0.18.3
gast==0.4.0
gdown==4.6.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
griffe==0.25.4
grpcio==1.51.3
h5py==3.8.0
hydra-core==1.3.1
idna==2.10
imageio==2.25.1
imbalanced-learn==0.10.1
importlib-metadata==6.1.0
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.10.0
ipywidgets==8.0.4
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-dash==0.4.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kaleido==0.2.1
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lightgbm==3.3.5
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocs-material==9.0.12
mkdocs-material-extensions==1.1.1
mkdocstrings==0.20.0
mkdocstrings-python==0.8.3
MouseInfo==0.1.3
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
numba==0.56.4
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
orjson==3.8.9
packaging==23.0
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly-resampler==0.8.3.2
pluggy==1.0.0
pmdarima==2.0.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycaret==3.0.0
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.0
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.6.0
pymdown-extensions==9.9.2
Pympler==1.0.1
PyMsgBox==1.0.9
pyod==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pyproject_hooks==1.0.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.1
pytest-cov==4.0.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywhatkit==5.4
pywin32==305
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rich==13.3.2
roboflow==0.2.29
rsa==4.9
schemdraw==0.16
scikit-image==0.19.3
scikit-plot==0.3.7
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.16.0
shapely==2.0.1
six==1.16.0
sktime==0.17.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.20.0
tb-nightly==2.13.0a20230213
tbats==1.1.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
thop==0.1.1.post2209072238
tifffile==2023.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.65.0
trace-updater==0.0.9
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
ultralytics==8.0.20
urllib3==1.26.14
utils==1.0.1
validators==0.20.0
vonage==3.2.2
watchdog==2.2.1
wcwidth==0.2.6
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.5
wikipedia==1.4.0
wrapt==1.14.1
xxhash==3.2.0
yacs==0.1.8
yapf==0.32.0
zipp==3.15.0
